requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
discord==0.0.309
discord.py==1.7.3
entrypoints==0.4
executing==0.8.2
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
notebook-as-pdf==0.5.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
pure-eval==0.2.2
pycparser==2.21
pyee==8.2.2
Pygments==2.11.2
pyparsing==3.0.7
PyPDF2==1.26.0
pyppeteer==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
Send2Trash==1.8.0
six==1.16.0
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0
